LM-1657 stellar withdrawal failed - Added fees configuration

+++ src/Lykke.Tools.Stellar/Commands/BalanceCalculationCommand.cs
@@ -88,7 +88,7 @@ namespace Lykke.Tools.Stellar.Commands
             {
                 try
                 {
-                    if (_latestBlock.HasValue && transaction.Ledger > _latestBlock)
+                    if (_latestBlock.HasValue && (long)transaction.Ledger > _latestBlock)
                     {
                         cursor = null;
                         break;
